Refactor FileProductStorage into BaseProductStorage, add S3ProductStorage, unused option to not copy contents
